tut5: layout needs to be flattened before extraction
looks like a zero-width fet was getting created in the extracted netlist
out of nowhere. seems like this is because of the tool not being able to
identify shapes that span multiple layers. needed to remove that fet.

post-layout sim for an inverter passed. layout for ro7 drc/lvs clean.
need to finish up sim for tut5.

Signed-off-by: Karthik B K <ee24s053@smail.iitm.ac.in>