tut5: simulation complete
bugs found and reported to srampr during the process of getting
these simulations to work:

1. zero-width fet generation in the pex-ed netlist because of not
   flattening the gds prior to extraction. seems like some funky shapes
were inferred as a fet because the extractor (magic) couldn't figure out
what they were. solution was to flatten the gds.

2. path picked up by magic-drc was messed up. this led to magic-drc not
   reporting any errors because it couldn't find the correct gds to look
at in the first place. looks like this was scripts written by us to integrate stuff.. the fix was to make a correction in those scripts to pick up the correct path.

Signed-off-by: Karthik B K <ee24s053@smail.iitm.ac.in>