mywork: nand2x1: drc/lvs clean and pex complete
when i made a contact from li1 to met2 over the DGND rail, it shorted
that net with DGND because li1-met2 goes through met1. moved the net
higher, as per suggestions from the prof, the layout is now drc/lvs
clean. diffusion sharing in the nmos can make the layout more compact,
maybe i can even switch the direction of source/drain in the nmos.

Signed-off-by: Karthik B K <ee24s053@smail.iitm.ac.in>